requirements.txt
streamlit
pandas
numpy
yfinance
python-bcb
plotly
pyotp
qrcode
mercadopago
google-auth-oauthlib
google-api-python-client
toml
tzlocal
tzdata
pandas-ta
scipy
requests
beautifulsoup4
